PBI-1: Move Login.js and Login.css into src/pages/auth, update imports
