+++ src/index.js
@@ -2,7 +2,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import Login from "./Login";
+import Login from './pages/auth/Login';
+
 
 import { PublicClientApplication } from "@azure/msal-browser";
 import { MsalProvider } from "@azure/msal-react";
